レイアウトと通知機能
