+++ app/controllers/public/favorites_controller.rb
@@ -1,13 +1,11 @@
 class Public::FavoritesController < ApplicationController
   before_action :check_customer, only: [:create, :destroy]
   def create
-    # @character= Character.find(params[:character_id])
      @comment = Comment.find(params[:comment_id])
      @favorite = current_customer.favorites.new(comment_id: @comment.id)
      @favorite.save
   end
   def destroy
-    # @character = Character.find(params[:character_id])
      @comment = Comment.find(params[:comment_id])
     @favorite = current_customer.favorites.find_by(comment_id: @comment.id)
     @favorite.destroy
