+++ app/javascript/packs/application.js
@@ -11,7 +11,8 @@ import "channels"
 import "jquery";
 import "popper.js";
 import "bootstrap";
-import "../stylesheets/application" 
+import "../stylesheets/application"
+import "../stylesheets/customer"
 import '@fortawesome/fontawesome-free/js/all'
 
 Rails.start()
