+++ app/javascript/stylesheets/application.scss
@@ -9,7 +9,6 @@ header {
 main{
   min-height: calc(100vh - 56px - 136px);
   background-color: #EEEEEE;
-  // padding-bottom: 100px;
 }
 
 footer {
