+++ app/javascript/stylesheets/customer.scss
@@ -1,0 +1,9 @@
+
+@media (min-width: 768px) {
+  .device-alert {
+    position: absolute;
+    top: 9px;
+    left: 5px;
+    font-size: 9px;
+  }
+}
