+++ app/views/layouts/_header.html.erb
@@ -38,12 +38,13 @@
             <i class="fa-solid fa-otter"></i>キャラクター一覧
           <% end %>
         </li>
-        <% if @unchecked_notifications.any? %>
-          <i class="fa-sharp fa-solid fa-circle"></i>
+        <li class="nav-item active" style="position:relative">
+                  <% if @unchecked_notifications.any? %>
+          <i class="fa-sharp fa-solid fa-circle device-alert" ></i>
         <% end %>
-        <li class="nav-item active">
+
           <%= link_to  admin_contacts_path, class: "nav-link" do %>
-            <i class="fa-regular fa-bell"></i>通知
+            <i class="fa-regular fa-bell" ></i>通知
           <% end %>
         </li>
         <li class="nav-item active">
