+++ app/views/public/contacts/done.html.erb
@@ -8,5 +8,5 @@
     <div class="col-5"><p class="h2">メッセージ</p></div>
     <div class="col-7"><p class="h4"><%= @contact.message %></p></div>
   </div>
-  <div class="text-center mt-4"><%= link_to "キャラクター一覧へ", admin_contacts_path(@contact), method: :post, class: "btn btn-secondary" %></div>
+  <div class="text-center mt-4"><%= link_to "キャラクター一覧へ", characters_path, class: "btn btn-secondary" %></div>
 </div>
